x64 fixes.

+++ Platform/DirectX11/Direct3D11Manager.cpp
@@ -343,7 +343,7 @@ void Direct3D11Manager::Initialize()
 	m_videoCardMemory = (int)(adapterDesc.DedicatedVideoMemory / 1024 / 1024);
 
 	// Convert the name of the video card to a character array and store it.
-	unsigned stringLength;
+	size_t stringLength;
 	error = wcstombs_s(&stringLength, m_videoCardDescription, 128, adapterDesc.Description, 128);
 	SIMUL_ASSERT(error==0);
 	//Now that we have stored the numerator and denominator for the refresh rate and the video card information we can release the structures and interfaces used to get that information.
